Algorithm_Assignment_5
Goal: input a square of grass and bush, and the range of landing. Output the biggest grass area.

Thoughts: BFS Search each area, if the up, down, left, and right square is grass, push it into the queue and turn it into bush (count once only)

Careful: It is still included if it's grass that we can't land on.